refactor(frontend): extract duplicated extractApiError to lib/utils
conceptos, documentos, tramites, estados-gestion, and folios pages each
defined their own copy of the 409-conflict-body parser (folios' copy
additionally lacked the [409] guard, silently attempting JSON-parse on
any error). Consolidated into a single lib/utils.ts implementation with
the [409] guard applied consistently, preserving existing behavior for
all five pages.

Closes #616
