+++ frontend/src/app/dashboard/administracion/documentos/page.tsx
@@ -12,22 +12,10 @@ import { Dialog, DialogContent } from "@/components/ui/dialog";
 import { FormContainer, FormSection, FormField, FormActions } from "@/theme/form-patterns";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { apiGet, apiPost, apiPut, apiDelete } from "@/lib/api-client";
+import { extractApiError } from "@/lib/utils";
 import type { TipoDeDocumento } from "@/types";
 
 const EMPTY: Partial<TipoDeDocumento> = { nombre: "" };
-
-function extractApiError(err: unknown): string | null {
-  if (!(err instanceof Error)) return null;
-  if (!err.message.includes("[409]")) return null;
-  try {
-    const jsonStart = err.message.indexOf("{");
-    if (jsonStart !== -1) {
-      const body = JSON.parse(err.message.slice(jsonStart)) as { error?: string };
-      return body.error ?? null;
-    }
-  } catch { }
-  return null;
-}
 
 export default function DocumentosPage() {
   const t = useTranslations("administracion.documentos");
